requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic 
python-multipart
pydantic-settings>=2.0.0
matplotlib
seaborn
numpy
psutil